Implement Versioning
Why?
The API consumer I SHOULD be able to continue the version I'm using without issues by changes made to the API

How?
Move all the controllers to `v1` folder
Enclose the controllers in `V1` module
Move the controllers specs to `v1` directory
Prefix all controller calls with the `V1::` namespace in the specs
Enclose the routes in the `v1` namespace

Completing https://www.pivotaltracker.com/story/show/111833517
[finished #111833517]
